chore: fully split events related modules from services package
