+++ app/infrastructure/events/__init__.py
@@ -5,9 +5,10 @@ cross-feature communication.
 """
 
 from infrastructure.events.models import Event
-from infrastructure.events.service import EventDispatcher
+from infrastructure.events.service import EventDispatcher, get_event_dispatcher
 
 __all__ = [
     "Event",
     "EventDispatcher",
+    "get_event_dispatcher",
 ]
