+++ app/infrastructure/services/__init__.py
@@ -6,15 +6,13 @@ Provides type aliases and provider functions for FastAPI dependency injection.
 
 from infrastructure.services.dependencies import (
     AppSettingsDep,
-    SettingsDep,
     MaxMindClientDep,
-    EventDispatcherDep,
+    SettingsDep,
 )
 from infrastructure.services.providers import (
     get_app_settings,
-    get_settings,
     get_maxmind_client,
-    get_event_dispatcher,
+    get_settings,
 )
 
 __all__ = [
@@ -22,10 +20,8 @@ __all__ = [
     "AppSettingsDep",
     "SettingsDep",
     "MaxMindClientDep",
-    "EventDispatcherDep",
     # Core providers
     "get_app_settings",
     "get_settings",
     "get_maxmind_client",
-    "get_event_dispatcher",
 ]
